fixing TOOLS-1612
git-svn-id: https://svn.wso2.org/repos/wso2/trunk/tools@169211 a5903396-d722-0410-b921-86c7d4935375

+++ ide/eclipse/esb/org.wso2.developerstudio.eclipse.gmf.esb.diagram/src/org/wso2/developerstudio/eclipse/gmf/esb/diagram/custom/deserializer/TaskDeserializer.java
@@ -79,7 +79,7 @@ public class TaskDeserializer extends AbstractEsbNodeDeserializer<TaskDescriptio
 			executeSetValueCommand(TASK__TRIGGER_TYPE, TaskTriggerType.CRON);
 		} else {
 			executeSetValueCommand(TASK__TRIGGER_TYPE, TaskTriggerType.SIMPLE);
-			executeSetValueCommand(TASK__COUNT, task.getCount());
+			executeSetValueCommand(TASK__COUNT, (long)task.getCount());
 			executeSetValueCommand(TASK__INTERVAL, task.getInterval() / 1000);
 		}
